+++ spec/geo_works/derivatives/runners/vector_derivatives_spec.rb
@@ -4,8 +4,6 @@ require 'fileutils'
 
 RSpec.describe GeoWorks::Derivatives::Runners::VectorDerivatives do
   describe "#create" do
-    let(:display_vector_uri) { test_derivative_url('display_vector', 'zip') }
-    let(:thumbnail_uri) { test_derivative_url('thumbnail', 'png') }
     let(:outputs) do
       [
         {
@@ -32,14 +30,18 @@ RSpec.describe GeoWorks::Derivatives::Runners::VectorDerivatives do
     end
 
     after do
-      # Cleanup generated derivatives
-      FileUtils.rm(display_vector_uri.path)
-      FileUtils.rm(thumbnail_uri.path)
+      # Cleanup generated derivatives, unless KEEP env variable is set
+      unless ENV['KEEP']
+        FileUtils.rm(display_vector_uri.path)
+        FileUtils.rm(thumbnail_uri.path)
+      end
     end
 
     context "with a zipped shapefile" do
       let(:input_file_path) { Pathname.new(test_data_fixture_path('files/vector/shapefile.zip')) }
       let(:input_mime_type) { 'application/zip; ogr-format="ESRI Shapefile"' }
+      let(:display_vector_uri) { test_derivative_url('shapefile_display_vector', 'zip') }
+      let(:thumbnail_uri) { test_derivative_url('shapefile_thumbnail', 'png') }
 
       it_behaves_like "a set of vector derivatives"
     end
@@ -47,6 +49,8 @@ RSpec.describe GeoWorks::Derivatives::Runners::VectorDerivatives do
     context "with a geojson file" do
       let(:input_file_path) { Pathname.new(test_data_fixture_path('files/vector/geojson.json')) }
       let(:input_mime_type) { 'application/vnd.geo+json' }
+      let(:display_vector_uri) { test_derivative_url('geojson_display_vector', 'zip') }
+      let(:thumbnail_uri) { test_derivative_url('geojson_thumbnail', 'png') }
 
       it_behaves_like "a set of vector derivatives"
     end
@@ -54,6 +58,8 @@ RSpec.describe GeoWorks::Derivatives::Runners::VectorDerivatives do
     context "with a KML file" do
       let(:input_file_path) { Pathname.new(test_data_fixture_path('files/vector/example.kml')) }
       let(:input_mime_type) { 'application/vnd.google-earth.kml+xml' }
+      let(:display_vector_uri) { test_derivative_url('kml_display_vector', 'zip') }
+      let(:thumbnail_uri) { test_derivative_url('kml_thumbnail', 'png') }
 
       it_behaves_like "a set of vector derivatives"
     end
